requirements.txt
fastapi
uvicorn
python-multipart
boto3
python-dotenv
pydantic
requests
PyJWT 